Renamed angulartics-google-analytics.js to angulartics-ga.js to avoid adblock blocking the source
