+++ example/index.html
@@ -6,7 +6,7 @@
 	<link rel="stylesheet" href="//bootswatch.com/cosmo/bootstrap.min.css">
 	<script src="//ajax.googleapis.com/ajax/libs/angularjs/1.1.5/angular.min.js"></script>
 	<script src="/node_modules/angulartics/dist/angulartics.min.js"></script>
-	<script src="/dist/angulartics-google-analytics.min.js"></script>
+	<script src="/dist/angulartics-ga.min.js"></script>
 	<script>
 	  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
 	  (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
