+++ lib/angulartics-ga.js
