+++ lib/index.js
@@ -1,2 +1,2 @@
-require('./angulartics-google-analytics')
+require('./angulartics-ga')
 module.exports = 'angulartics.google.analytics'
